# target/classes/openapi.yaml
openapi: 3.0.3
info:
  title: DigiShopKart
  description: Apis of E-commerce application
  contact:
    name: Ajay Kate
    url: http://digishopkart.com
  version: '0.1'
servers:
  - url: http://localhost:8085
#------------------------------PATHS ---------------------------------
paths:
...
  /api/v1/discounts :
    description: Api to create discount coupon
    post:
      tags:
        - Discount
      summary: fetch a discount
      operationId: createDiscount
      description: create Discount
      requestBody:
        $ref: '#/components/requestBodies/DiscountBody'
      responses:
        200:
          description: sucess response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Discount'
        404:
          $ref: '#/components/responses/NotFound'
        400:
          $ref: '#/components/responses/BadRequest'
    get:
      tags:
        - Discount
      summary: fetch all discounts
      operationId: fetchAllDiscounts
      description: Api to fetch All the Discount coupons
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Discount'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'
  /api/v1/discounts/{discountId}:
    description: APi to fetch Discount Coupon by Id
    get:
      tags:
        - Discount
      operationId: fetchDiscount
      summary: fetch a discount
      description: getById api
      parameters:
        - in: path
          name: discountId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: sucess response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Discount'
        404:
          $ref: '#/components/responses/NotFound'
        400:
          $ref: '#/components/responses/BadRequest'

    put:
      tags:
        - Discount
      summary: update a discount
      operationId: updateDiscount
      description: Api to update Discount Coupons recors
      parameters:
        - in: path
          name: discountId
          required: true
          schema:
            type: integer
      requestBody:
        $ref: '#/components/requestBodies/DiscountBody'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Discount'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - Discount
      summary: delete a discount
      operationId: deleteDiscount
      description: Api to delete Discount Coupon By id
      parameters:
        - in: path
          name: discountId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Discount'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'


  #  ********************************** SUBSCRIPTION  **********************************
...
  /api/v1/orders:
    post:
      tags:
        - Order
      summary: create an order
      operationId: createOrder
      description: Api to create Order by using Customer, Product and Discount
      requestBody:
        $ref: '#/components/requestBodies/OrderBody'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'

    get:
      tags:
        - Order
      summary: fetch all orders
      operationId: fetchAllOrders
      description: APi to get All the Order
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'
  /api/v1/orders/{orderId}:
    get:
      tags:
        - Order
      summary: fetch an order
      description: Api to fetch Order by id
      operationId: fetchOrder
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - Order
      summary: detete an order
      operationId: deleteOrder
      description: Api to delete the Order by id
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Order
      summary: update an order
      operationId: updateOrder
      description: Api to update Order record by id
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
      requestBody:
        $ref: '#/components/requestBodies/OrderBody'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags:
        - Order
      summary: update Order Status
      operationId: updateOrderStatus
      description: API to update the status of order
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
      requestBody:
        $ref: '#/components/requestBodies/OrderBody'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'
  #   .   .   .   .   .   .   .   .   .   .   .   .   .   .   .

#___________________RRESPONSES
components:
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
  #___________________REQUEST BODY_____________________________________
  requestBodies:
...
    DiscountBody:
      description: A JSON object Contenting Discount information.
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Discount'
...
    OrderBody:
      description: A JSON object Contenting Order information
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Order'


  #___________________SCHEMAS
  schemas:
    Errors:
      type: object
      properties:
        code:
          type: string
        message:
          title: string
      required:
        - code
        - message
    Customer:
      type: object
      required:
        - firstName
        - lastName
        - email
        - mobile
        - customerStatus
        - customerAddress
      properties:
        customerId:
          type: integer
        firstName:
          type: string
          example: Ajay
          pattern: '^[a-zA-Z]+(\s[a-zA-Z]+)?$'
        lastName:
          type: string
          example: Kate
          pattern: '^[a-zA-Z]+(\s[a-zA-Z]+)?$'
        email:
          type: string
          example: ajaykate@gmail.com
          format: email
        mobile:
          type: string
          example: +919876543210
        customerStatus:
          type: string
          enum:
            - active
            - inactive
          default: active
          nullable: false
        customerAddress:
          type: array
          items:
            $ref: '#/components/schemas/CustomerAddress'
    CustomerAddress:
      type: object
      required:
        - customerFullName
        - country
        - state
        - town
        - area
        - houseOrBuilding
        - landmark
        - pinCode
        - mobile
      properties:
        id:
          type: integer
        addressType:
          type: string
          enum:
            - shipping
            - billing
        customerFullName:
          type: string
        country:
          type: string
        state:
          type: string
        town:
          type: string
        area:
          type: string
        houseOrBuilding:
          type: string
        landmark:
          type: string
        pinCode:
          type: integer
        mobile:
          type: string
    Product:
      type: object
      required:
        - productName
        - createdAt
        - productCode
        - productPrice
        - productStatus
        - productCategory
      properties:
        productId:
          type: integer
        productName:
          type: string
        productCode:
          type: string
        productPrice:
          type: number
          minimum: 1
          format: double
        productCategory:
          type: string
          enum:
            - electronics
            - cloths
            - footwares
            - jwelleries
            - gadgets
        productStatus:
          type: string
          enum:
            - active
            - inactive
        variant:
          type: array
          items:
            $ref: '#/components/schemas/Variant'
    Variant:
      type: object
      required:
        - colour
        - size
      properties:
        id:
          type: integer
        name:
          type: string
          example: Size, Colour, etc...
        value:
          type: string
        status :
          type: string
          enum:
            - active
            - inactive
    Discount:
      type: object
      required:
        - discountType
        - couponValue
      properties:
        discountId:
          type: integer
        couponName:
          type: string
        discountType:
          type: string
          enum:
            - FLAT
            - PERCENTAGE
        couponValue:
          type: number
          format: double
...
    Order:
      type: object
      required:
        - orderId
        - activatedAt
        - productName
        - productPrice
        - couponValue
        - customer
        - product
        - discount
      properties:
        orderId:
          type: integer
        productName:
          type: string
        activatedAt:
          type: string
        discountPrice:
          type: number
          minimum: 1
          format: double
        couponValue:
          type: number
          format: double
        orderStatus:
          type: string
          enum:
            - confirmed
            - placed
            - dispatched
            - outfordelivery
            - delivered
            - cancelled
        customer:
          $ref: '#/components/schemas/Customer'
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        discount:
          $ref: '#/components/schemas/Discount'

    #---------------------------------------------------------------------------